Each class is interacting properly
I went through and made a few methods that checked to see if a ticket would be issued, I created some static methods so I wouldn't have to pass so many parameters via the main function, and now it seems that information is flowing properly. I next have to fix my math to properly calculate fines, because as of right now it is calculating the fines incorrectly. I also have a few duplicates that I need to get rid of, but I will focus on cleaning up the code after I fix the math in the calculateFine() method.