Used inheritance to lessen the amount of variables declared
I made sure that ParkingTicket has access to methods and variables in ParkedCar to rid my code of excessive variables. I Also fixed the math in the calculateFine method so that being only 1 minute to 60 minutes over would give a $25 fine only. For every 1-60 minutes after the first hour then the extra $10 is added based upon how many hours it is over the first hour over. In the main function I just created a ParkingTicket object and passed it into the issueTicket() method which checks if a ticket will or will not be issued.