Created new objects for each class
So I passed each class into the Parking Ticket Class itself so it would be a lot easier to understand where information would get lost or corrupted, or if a class specifically was not functioning correctly. I do believe this was a better choice and makes my code a lot easier to document and debug.